implement http redirect in the library.
This is a pretty quick implementation. I didn't want to refactor the world, but it gets the job done.
Unfortunately theres no way in the current class layout to fudge the host to this new host one time.
I set the instance host to the new redirect, which should be leader. This is pretty gnarly. Perhaps
creating a new client instance and setting that up with the new host would be a better behavior.