Add simple specs for `eternal_watch`
Before this method can become more sophisticated, it should be tested
a little. In these two specs, we exercise the basic
behaviour (watching for multiple changes to a key, instead of
returning after one) and that recursive keys can be watched.

The `eternal_watch` method uses `loop`, which would normally be hard
to test, but we can control it in the specs by breaking out of the
loop using an exception once we know we've exercised everything we
care about. I've added some code to help manage that in a simple way.