Add failing test for resuming watching after multiple updates
It's good practice for etcd watchers to track what the latest update
they received was, and then resume watching immediately after that
update, to ensure that no changes might've been missed between the
return of the first watch and the start of the second; i.e. when
actual work on the changed key/value is being done.

At the moment, `eternal_watch` doesn't do that, so if the yielded
block doesn't return before the next key update, that update will be
lost.

Subsequent commits should fix this test and make the behaviour of
`eternal_watch` more robust.